implementation of edge-tts and abstract confiruration for trasformation to google tts
