+++ Backend/app/core/celery_app.py
@@ -1,7 +1,3 @@
-import asyncio
-import os
-import time
-
 from celery import Celery
 from celery.schedules import crontab
 
@@ -23,6 +19,7 @@ celery_app.conf.update(
         "workers.scraping_worker.scrape":       {"queue": "scraping"},
         "workers.filtering_worker.filter_pois": {"queue": "filtering"},
         "workers.narrative_worker.narrate":     {"queue": "narrative"},
+        "workers.cleanup_worker.cleanup_audio": {"queue": "cleanup"},
     },
 )
 
@@ -34,38 +31,5 @@ celery_app.conf.beat_schedule = {
 }
 
 
-@celery_app.task(name="workers.narrative_worker.narrate", bind=True)
-def narrate(self, text: str, language: str, speed: int, pitch: int, loudness: int) -> dict:
-    from app.core.config import settings as _settings
-    from app.services.tts.factory import TTSFactory
-
-    os.makedirs(_settings.AUDIO_DIR, exist_ok=True)
-    provider = TTSFactory.get_provider()
-    audio_bytes = asyncio.run(
-        provider.synthesize(text, language, speed=speed, pitch=pitch, loudness=loudness)
-    )
-    filename = f"{self.request.id}.mp3"
-    filepath = os.path.join(_settings.AUDIO_DIR, filename)
-    with open(filepath, "wb") as f:
-        f.write(audio_bytes)
-    return {"audio_url": f"/audio/{filename}"}
 
 
-@celery_app.task(name="workers.cleanup_worker.cleanup_audio")
-def cleanup_audio() -> dict:
-    from app.core.config import settings as _settings
-
-    audio_dir = _settings.AUDIO_DIR
-    if not os.path.exists(audio_dir):
-        return {"deleted": 0}
-
-    now = time.time()
-    deleted = 0
-    for fname in os.listdir(audio_dir):
-        if fname == ".gitkeep":
-            continue
-        fpath = os.path.join(audio_dir, fname)
-        if os.path.isfile(fpath) and (now - os.path.getmtime(fpath)) > settings.AUDIO_FILE_TTL_SECONDS:
-            os.remove(fpath)
-            deleted += 1
-    return {"deleted": deleted}
