+++ Backend/tests/workers/test_cleanup_worker.py
@@ -1,9 +1,8 @@
 import os
 import time
 import pytest
-
-import app.core.celery_app as celery_app_module
-
+from app.core.config import settings
+from workers.cleanup_worker import cleanup_audio
 
 def _create_file(directory: str, name: str, age_seconds: float) -> str:
     path = os.path.join(directory, name)
@@ -15,13 +14,13 @@ def _create_file(directory: str, name: str, age_seconds: float) -> str:
 
 
 def test_cleanup_deletes_old_files(tmp_path, monkeypatch):
-    monkeypatch.setattr("app.core.config.settings.AUDIO_DIR", str(tmp_path))
-    monkeypatch.setattr("app.core.config.settings.AUDIO_FILE_TTL_SECONDS", 3600)
+    monkeypatch.setattr("workers.cleanup_worker.settings.AUDIO_DIR", str(tmp_path))
+    monkeypatch.setattr("workers.cleanup_worker.settings.AUDIO_FILE_TTL_SECONDS", 3600)
 
     old_file = _create_file(str(tmp_path), "old.mp3", age_seconds=7200)
     new_file = _create_file(str(tmp_path), "new.mp3", age_seconds=60)
 
-    result = celery_app_module.cleanup_audio.run()
+    result = cleanup_audio()
 
     assert result == {"deleted": 1}
     assert not os.path.exists(old_file)
@@ -29,12 +28,12 @@ def test_cleanup_deletes_old_files(tmp_path, monkeypatch):
 
 
 def test_cleanup_ignores_gitkeep(tmp_path, monkeypatch):
-    monkeypatch.setattr("app.core.config.settings.AUDIO_DIR", str(tmp_path))
-    monkeypatch.setattr("app.core.config.settings.AUDIO_FILE_TTL_SECONDS", 3600)
+    monkeypatch.setattr("workers.cleanup_worker.settings.AUDIO_DIR", str(tmp_path))
+    monkeypatch.setattr("workers.cleanup_worker.settings.AUDIO_FILE_TTL_SECONDS", 3600)
 
     _create_file(str(tmp_path), ".gitkeep", age_seconds=99999)
 
-    result = celery_app_module.cleanup_audio.run()
+    result = cleanup_audio()
 
     assert result == {"deleted": 0}
     assert os.path.exists(os.path.join(str(tmp_path), ".gitkeep"))
@@ -42,8 +41,8 @@ def test_cleanup_ignores_gitkeep(tmp_path, monkeypatch):
 
 def test_cleanup_no_directory(tmp_path, monkeypatch):
     missing_dir = str(tmp_path / "nonexistent")
-    monkeypatch.setattr("app.core.config.settings.AUDIO_DIR", missing_dir)
+    monkeypatch.setattr("workers.cleanup_worker.settings.AUDIO_DIR", missing_dir)
 
-    result = celery_app_module.cleanup_audio.run()
+    result = cleanup_audio()
 
     assert result == {"deleted": 0}
